requirements.txt
opencv-python~=4.10.0.84
numpy~=2.0.1
setuptools~=69.0.3
matplotlib~=3.8.2
scikit-image~=0.22.0
psycopg2~=2.9.9
dlib~=19.24.2
Pillow~=10.1.0
python-dotenv~=1.0.1
cmake~=3.30.2